+++ src/main/java/pl/pawkrol/academic/IssuesManager/project/ProjectController.java
@@ -1,9 +1,13 @@
 package pl.pawkrol.academic.IssuesManager.project;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import pl.pawkrol.academic.IssuesManager.session.SessionService;
+import pl.pawkrol.academic.IssuesManager.user.User;
 
+import javax.websocket.server.PathParam;
 import java.util.List;
 
 @RestController
@@ -11,10 +15,12 @@ import java.util.List;
 public class ProjectController {
 
     private ProjectService projectService;
+    private SessionService sessionService;
 
     @Autowired
-    public ProjectController(ProjectService projectService) {
+    public ProjectController(ProjectService projectService, SessionService sessionService) {
         this.projectService = projectService;
+        this.sessionService = sessionService;
     }
 
     @PutMapping("/save")
@@ -34,7 +40,31 @@ public class ProjectController {
 
     @DeleteMapping("/remove")
     ResponseEntity remove(@RequestParam String id) {
+        Project project = projectService.getById(id);
+        if (!project.getOwnerId().equals(sessionService.getUserId())) {
+            return ResponseEntity
+                    .status(HttpStatus.FORBIDDEN)
+                    .build();
+        }
+
         projectService.remove(id);
-        return ResponseEntity.ok().build();
+        return ResponseEntity
+                .ok()
+                .build();
+    }
+
+    @PostMapping("/add-user/{projectId}")
+    ResponseEntity addUserToProject(@PathVariable String projectId,
+                                    @PathParam(value = "username") String username) {
+        if (projectService.addUserToProject(projectId, username)) {
+            return ResponseEntity.ok().build();
+        }
+
+        return ResponseEntity.notFound().build();
+    }
+
+    @GetMapping("/users/{projectId}")
+    List<User> getProjectUsers(@PathVariable String projectId) {
+        return projectService.getProjectUsers(projectId);
     }
 }
